abc123/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
mysqlclient>=2.1
django-cors-headers>=4.7